move core functions into src folder
