+++ sequencers/randSeqPerFrame.js
