+++ src/compare-frames.js
