+++ src/sequitur.js
@@ -28,10 +28,10 @@ if (!fs.existsSync(dir))
     fs.mkdirSync(dir);
 
 // load data
-exports.frameCount = function () { return require("./extractFrames").frames(video, size); }
+exports.frameCount = function () { return require("./extract-frames").frames(video, size); }
 exports.diffs = function () { 
-        const frameCount = require("./extractFrames").frames(video, size);
-        return require('./compareFrames').diffs(frameCount);
+        const frameCount = require("./extract-frames").frames(video, size);
+        return require('./compare-frames').diffs(frameCount);
 }
-exports.wave = function () { return require('./sampleAudio').wave(audio, framerate); }
-exports.export = function (sequence) { require('./exportSequence').concat(sequence, size, framerate, audio, preview); }
+exports.wave = function () { return require('./sample-audio').wave(audio, framerate); }
+exports.export = function (sequence) { require('./export-sequence').concat(sequence, size, framerate, audio, preview); }
